move queries to external files
