+++ src/containers/QuestionnaireCreatePage/index.js
@@ -1,13 +1,6 @@
-import { graphql, gql } from "react-apollo";
+import { graphql } from "react-apollo";
 import QuestionnaireCreate from "./QuestionnaireCreate";
-
-export const createQuestionnaire = gql`
-  mutation createQuestionnaire($title: String!, $description: String!, $theme: String!, $legalBasis: String!, $navigation: Boolean) {
-    createQuestionnaire(title: $title, description: $description, theme: $theme, legalBasis: $legalBasis, navigation: $navigation) {
-      id
-    }
-  }
-`;
+import createQuestionnaire from "queries/createQuestionnaire";
 
 export const withMutation = graphql(createQuestionnaire, {
   props: ({ ownProps, mutate }) => ({
